refactored node in tree too
