+++ data-structures/trees/node.js
@@ -1,0 +1,9 @@
+class Node {
+  constructor(val) {
+    this.val = val;
+    this.left = null;
+    this.right = null;
+  }
+}
+
+module.exports = Node;
